crewai/requirements.txt
# HyperMemory CrewAI Integration Requirements

crewai>=0.1.0
crewai-tools>=0.1.0
requests>=2.28.0
python-dotenv>=1.0.0
